+++ .github/workflows/ci.yml
@@ -185,8 +185,8 @@ jobs:
     - run: yarn null:check
       timeout-minutes: 5
 
-    - name: check null autoadd
-      run: yarn null:autoadd && node verify-clean-strict-null-checks.js
+    - name: verify null:autoadd finds no changes
+      run: yarn null:autoadd && node verify-unchanged-strict-null-checks.js
       timeout-minutes: 6
 
   codeql:
